Refactor XML classes
Use PIMPL class in the XML classes and put implementation in the source
files instead of in a different class (it was overcomplicated).
Put both classes (XmlNode and XmlDocument) in the same files (header and
source). This is needed because the document needs access to the node
implementation to be able to construct nodes. TODO: Find a way to
separate both classes in their specific files (one first level class per
header rule).
Implement XmlDocument::getRootNode().
Implement XmlNode::getNAme().